+++ hutool-core/src/main/java/cn/hutool/core/bean/copier/MapToBeanCopier.java
@@ -2,7 +2,6 @@ package cn.hutool.core.bean.copier;
 
 import cn.hutool.core.bean.BeanUtil;
 import cn.hutool.core.bean.PropDesc;
-import cn.hutool.core.convert.Convert;
 import cn.hutool.core.lang.Assert;
 import cn.hutool.core.map.CaseInsensitiveMap;
 import cn.hutool.core.map.MapWrapper;
@@ -83,7 +82,8 @@ public class MapToBeanCopier<T> extends AbsCopier<Map<?, ?>, T> {
 
 			// 获取目标字段真实类型并转换源值
 			final Type fieldType = TypeUtil.getActualType(this.targetType, tDesc.getFieldType());
-			Object newValue = Convert.convertWithCheck(fieldType, sValue, null, this.copyOptions.ignoreError);
+			//Object newValue = Convert.convertWithCheck(fieldType, sValue, null, this.copyOptions.ignoreError);
+			Object newValue = this.copyOptions.convertField(fieldType, sValue);
 			newValue = copyOptions.editFieldValue(sKeyStr, newValue);
 
 			// 目标赋值